whole test

+++ detect_BCode/include/find_ball.h
@@ -12,7 +12,7 @@ namespace RedContritio
 
 struct BallInfo
 {
-    cv::Point2d center;
+    cv::Point2f center;
     double radius;
 };
